move display_costs function to layout elements module
